Refactor bundle-size to split API methods from GitHub webhook methods
